@@ -5,9 +5,9 @@
   "main": "index.js",
   "private": true,
   "scripts": {
-    "start": "node index.js",
-    "lint": "eslint ./src",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "dev": "nodemon -x 'node -r esm' index.js",
+    "start": "node -r esm index.js",
+    "lint": "eslint ./src"
   },
   "license": "MIT",
   "engines": {
@@ -15,9 +15,10 @@
     "npm": ">=5.0.0"
   },
   "devDependencies": {
-    "eslint": "^4.19.1",
-    "eslint-config-airbnb-base": "^13.0.0",
-    "eslint-plugin-import": "^2.13.0"
+    "eslint": "^5.9.0",
+    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-plugin-import": "^2.14.0",
+    "nodemon": "^1.18.6"
   },
   "dependencies": {
     "bluebird": "^3.5.1",
@@ -26,11 +27,12 @@
     "cookie-parser": "^1.4.3",
     "crypto-js": "^3.1.9-1",
     "csurf": "^1.9.0",
-    "debug": "^3.1.0",
-    "envalid": "^4.1.1",
-    "express": "^4.16.2",
+    "debug": "^4.1.0",
+    "envalid": "^4.1.4",
+    "esm": "^3.0.84",
+    "express": "^4.16.4",
     "express-session": "^1.15.6",
-    "http-proxy-middleware": "^0.17.4",
+    "http-proxy-middleware": "^0.19.0",
     "redis": "^2.8.0",
     "zipkin": "^0.10.1",
     "zipkin-context-cls": "^0.6.1",
