@@ -1,11 +1,13 @@
-const csrf = require('csurf')();
+import csurf from 'csurf';
+
+const csrf = csurf();
 
 /**
  * @param {Request} req
  * @param {Response} res
  * @param {Function} next
  */
-module.exports = (req, res, next) => {
+export default (req, res, next) => {
     // Skip CSRF on requests with explicit Authorization header
     if (req.header('authorization')) {
         next();
