@@ -1,5 +1,6 @@
-/*
- * Launch proxy server.
- */
+import app from './src/app';
+import { HOST, PORT } from './src/env';
 
-require('./src/listen')();
+app.listen(PORT, HOST, () => {
+    process.stdout.write(`Proxy listening on ${HOST}:${PORT}`);
+});
